refactor(tests): use public API imports in execution tests
Replace deep imports (vibe3.clients.sqlite_client, vibe3.models.X) with
public API paths (vibe3.clients, vibe3.models) across 12 execution test
files. Add missing public exports init_schema and fail_issue.

Changes:
- src/vibe3/clients/__init__.py: add init_schema export
- src/vibe3/services/__init__.py: add fail_issue export
- test_command_adapter.py: 6 inline imports from vibe3.roles.X
- test_codeagent_runner_gate.py: vibe3.agents.models → vibe3.agents
- test_contract_compat.py: 4 imports from vibe3.models.X
- test_issue_role_support.py: 2 deep imports
- test_issue_role_sync_runner.py: vibe3.roles.definitions → vibe3.roles
- test_execution_role_policy.py: 5 imports from vibe3.models.X
- test_noop_gate_unit.py: 4 inline imports
- test_noop_gate_single_step_limit.py: vibe3.clients.sqlite_schema
- test_store_context.py: vibe3.clients.sqlite_client
- test_resume_resets_transition_count.py: 3 deep imports
- test_retry_counter_persistence.py: 2 deep imports
- test_error_block_decoupling.py: 3 deep imports

Tests: All 224 execution tests pass

Closes #2469
