+++ src/vibe3/services/__init__.py
@@ -34,6 +34,7 @@ if TYPE_CHECKING:
     from vibe3.services.issue.failure import (
         block_manager_noop_issue,
         fail_executor_issue,
+        fail_issue,
         fail_manager_issue,
         fail_planner_issue,
         fail_reviewer_issue,
@@ -136,6 +137,7 @@ __all__ = [
     "create_flow_manager",
     "emit_issue_failed",
     "fail_executor_issue",
+    "fail_issue",
     "fail_manager_issue",
     "fail_planner_issue",
     "fail_reviewer_issue",
@@ -239,6 +241,7 @@ _SYMBOL_MODULES = {
     "create_flow_manager": "vibe3.services.flow_factory",
     "emit_issue_failed": "vibe3.services.event_helpers",
     "fail_executor_issue": "vibe3.services.issue.failure",
+    "fail_issue": "vibe3.services.issue.failure",
     "fail_manager_issue": "vibe3.services.issue.failure",
     "fail_planner_issue": "vibe3.services.issue.failure",
     "fail_reviewer_issue": "vibe3.services.issue.failure",
