+++ tests/vibe3/execution/test_error_block_decoupling.py
@@ -13,12 +13,10 @@ from unittest.mock import patch
 
 import pytest
 
-from vibe3.clients.sqlite_client import SQLiteClient
-from vibe3.exceptions.runtime_errors import GitHubAPIError
+from vibe3.clients import SQLiteClient
+from vibe3.exceptions import GitHubAPIError
 from vibe3.execution.noop_gate import apply_unified_noop_gate
-from vibe3.services.issue.failure import (
-    fail_issue,
-)
+from vibe3.services import fail_issue
 
 
 @pytest.fixture
