+++ tests/vibe3/execution/test_resume_resets_transition_count.py
@@ -9,10 +9,10 @@ This test verifies the fix for issue #1880:
 import sqlite3
 from unittest.mock import MagicMock, patch
 
-from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.clients import SQLiteClient
 from vibe3.execution.noop_gate import apply_unified_noop_gate
-from vibe3.models.orchestration import IssueState
-from vibe3.services.blocked_state_service import BlockedStateService
+from vibe3.models import IssueState
+from vibe3.services import BlockedStateService
 
 
 def test_resume_resets_single_step_limit_counter(tmp_path):
