+++ tests/vibe3/execution/test_retry_counter_persistence.py
@@ -12,8 +12,8 @@ from unittest.mock import patch
 
 import pytest
 
-from vibe3.clients.sqlite_client import SQLiteClient
-from vibe3.exceptions.runtime_errors import GitHubAPIError
+from vibe3.clients import SQLiteClient
+from vibe3.exceptions import GitHubAPIError
 from vibe3.execution.noop_gate import apply_unified_noop_gate
